made changes to submit project
